requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
torch>=1.10.0
torchvision>=0.11.0

# Image processing
opencv-python>=4.5.0
scikit-image>=0.18.0
Pillow>=8.0.0

# Data handling
xmltodict>=0.12.0
tqdm>=4.60.0

# Visualization
matplotlib>=3.4.0

# Training & logging
wandb>=0.12.0

# Development (optional)
# pytest>=6.0.0
# black>=21.0
# flake8>=3.9.0
